split unfold op interface into separate file
